requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
pandas==1.5.3
selenium==4.21.0
webdriver-manager==4.0.1
Pillow==9.2.0

